Add ValueCodeGenerator Objects and Records support
Add support for generating code to create common value Objects and
Records. A common value Record is a public, accessible Record class
that has common value objects for all of its components. A common
value Object is a public, accessible Object class that has public
setters for all of its non-public instance fields, with each field
being assigned a common value.

The code generated for a common value Record calls the Record's
canonical constructor with the generated code for each of its
component. For instance, for `record MyRecord(String a, int b)`,
and instance `MyRecord("Example", 1)`, it would generate the
following code:

    new MyRecord("Example", 1)

The code generated for a common value Object calls the Object's
public no-args constructor and sets its fields. Public fields are
set by direct access. Non-public fields are set by calling their
setters. The code will be wrapped by an inlined Supplier so the
generated code can be used anywhere. For example, given the
following class:

    class MyObject {
        public String name;
        private int age;

        public void setAge(int age) {
            this.age = age;
        }
    }

The code generated for MyObject("Example", 2) would be

    ((Supplier<MyObject>) () -> {
        MyObject $valueCodeGeneratorObject1 = new MyObject();
        $valueCodeGeneratorObject1.setAge(2);
        $valueCodeGeneratorObject1.name = "Example";
        return $valueCodeGeneratorObject1;
    }).get()

ValueCodeGenerator uses a ThreadLocal keeping track of
the number of generateCode calls in the current Thread's stack.
This allows the Delegate for Object to use an unused identifier
for the variable holding the Object under construction.

Known limitations are that cyclic references cannot be handled.
If a cyclic reference is detected, a ValueCodeGenerationException
will be raised.

Additionally, if the same Object is used for multiple
fields/components in Objects/Records, they will be different
instances instead of the same instance.

This change will allow common value objects and records to be
passed to the constructor of BeanDefinitions in an AOT native image.
For instance, the following will now work in native image, provided
`myConfig` is a common value object:

    MyConfig myConfig = new MyConfig();
    // ...Set fields of MyConfig ...
    RootBeanDefinition rootBeanDefinition = new RootBeanDefinition(
        MyBean.class);
    rootBeanDefinition.setFactoryMethodName("of");
    rootBeanDefinition.getConstructorArgumentValues()
       .addGenericArgumentValue(myConfig);
    registry.registerBeanDefinition(BEAN_NAME,
        rootBeanDefinition);