Make BodyBuilder.bodyValueAndAwait() a reified type of function
Function `bodyValueAndAwait(body: Any)` is not a reified function and is
subject to type erasure. If an application uses for serialization the
`kotlinx.serialization` library and the provided body uses generics,
then the library is unable to serialize the body and spring falls back
to default one. This could lead to unexpected behaviors.

Signed-off-by: George Papadopoulos <george.719pap@gmail.com>